backend/pyproject.toml
[project]
name = "todo-console"
version = "0.1.0"
description = "Phase I: In-Memory Python Console Todo App - Hackathon II"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.132.0",
    "openai>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
]

[project.scripts]
todo = "src.main:main"

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "agents"]
